feat: wire p2p across trip document
